+++ src/Metrics/SSMetrics/Regions/OverlapBlock.cpp
@@ -3,13 +3,13 @@
 using namespace std;
 
 OverlapBlock::OverlapBlock(const SSBlock& refRegion, const SSBlock& predRegion) {
-    int predFrom = predRegion.GetFrom();
-    int predTo = predRegion.GetTo();
-    int refFrom = refRegion.GetFrom();
-    int refTo = refRegion.GetTo();
+    int64_t predFrom = predRegion.GetFrom();
+    int64_t predTo = predRegion.GetTo();
+    int64_t refFrom = refRegion.GetFrom();
+    int64_t refTo = refRegion.GetTo();
 
-    int overlapFrom = min(refFrom, predFrom);
-    int overlapTo = max(refTo, predTo);
+    int64_t overlapFrom = min(refFrom, predFrom);
+    int64_t overlapTo = max(refTo, predTo);
     this->_length = overlapTo - overlapFrom + 1;
 
     SetFrom(overlapFrom);
@@ -18,6 +18,6 @@ OverlapBlock::OverlapBlock(const SSBlock& refRegion, const SSBlock& predRegion) 
     this->predRegion = predRegion;
 }
 
-int OverlapBlock::GetLength() const {
+int64_t OverlapBlock::GetLength() const {
     return _length;
 }
