+++ src/Metrics/SSMetrics/Regions/Region.cpp
@@ -2,11 +2,11 @@
 
 using namespace std;
 
-int Region::GetTo() const {
+int64_t Region::GetTo() const {
     return _to;
 }
 
-void Region::SetTo(const int& to) {
+void Region::SetTo(const int64_t& to) {
     if (to < _from) {
         throw runtime_error("Region 'To' cannot be less than 'From'");
     }
@@ -14,16 +14,16 @@ void Region::SetTo(const int& to) {
     _length = _to - _from + 1;
 }
 
-int Region::GetFrom() const {
+int64_t Region::GetFrom() const {
     return _from;
 }
 
-void Region::SetFrom(const int& from) {
+void Region::SetFrom(const int64_t& from) {
     if (_from == -1) {
         _from = from;
     }
 }
 
-int Region::GetLength() const{
+int64_t Region::GetLength() const{
     return _length;
 }
