+++ src/Metrics/SSMetrics/include/Interfaces/INormMetric.hpp
@@ -7,14 +7,14 @@
 class INormMetric : public Metric
 {
 private:
-    int _nSum{0};
-    std::unordered_map<char, int> _normalizationSSMap;
+    int64_t _nSum{0};
+    std::unordered_map<char, int64_t> _normalizationSSMap;
 
-    int N(const char& secondaryStructure);
+    int64_t N(const char& secondaryStructure);
 public:
     INormMetric(const string& name, const string& refSequence, const string& predSequence, const bool& calculateNorm, PrecalculatedMetric* precalculated);
     virtual ~INormMetric() = default;
 
-    int _GetNormalizationSum() const;
-    int _GetNormalization(const char& secondaryStructure);  
+    int64_t _GetNormalizationSum() const;
+    int64_t _GetNormalization(const char& secondaryStructure);
 };
