+++ src/Metrics/SSMetrics/include/Interfaces/NormMetric.hpp
@@ -8,8 +8,8 @@ class NormMetric : public INormMetric
 private:
     PrecalculatedNormMetric* normPrecalculations;
 public:
-    int GetNormalizationSum() const;
-    int GetNormalization(const char& secondaryStructure);  
+    int64_t GetNormalizationSum() const;
+    int64_t GetNormalization(const char& secondaryStructure);
 
     NormMetric(const string& name, const string& refSequence, const string& predSequence, PrecalculatedNormMetric* precalculatedNorm, PrecalculatedMetric* precalculated);
     virtual ~NormMetric() = default;
