+++ src/Metrics/SSMetrics/include/StrictOverlap.hpp
@@ -9,7 +9,7 @@ private:
 
     bool GetZeroDelta() const;
     double DeltaSov(const OverlapBlock& overlapBlock);
-    int DeltaStrict(const OverlapBlock& overlapBlock);
+    int64_t DeltaStrict(const OverlapBlock& overlapBlock);
     double Theta(const OverlapBlock& overlapBlock);
 public:
     StrictOverlap(const string& name, const string& refSequence, const string& predSequence, const bool& zeroDelta, PrecalculatedMetric* precalculated);
